Notification component and sticky sidebar
Remove Phoenix Framework title suffix in favor of Beatseek.
User settings and login/registration working by setting the necessary `active_tab`. This may require assigns vs temporary assigns.
Add logout link which doesn't pass the call as a DELETE for some reason.
Add copyright footer with no styling. I can't make this sticky though. My hope was the header, sidebar, and footer would be sticky with scrolling in just the dynamic LiveView "page" sections.
Add notification badge component stubbing. I honestly don't get why I seemingly needed the `update` function to change the data. I thought doing that in the `mount` was enough. Components are just different enough apparently.

Made the sidebar sticky by removing the parent flex declaration. According to `https://www.designcise.com/web/tutorial/how-to-fix-issues-with-css-position-sticky-not-working#checking-if-a-parent-element-is-a-flexbox`, align-self seems to be a solution but I don't see a tailwind utility for it. We don't technically need that though as we could easily reuse a class.