Documentation updates
I mention being harsh in my criticism of `phx.gen.auth` when I'm in such a different mindset. I think having some wins with a sticky sidebar and a finished album screen has loosened my crabbiness. I also got a lot better sleep in the last 2 days in probably 2+ months so that probably has a LOT to do with it. Also the beta cohort's last instructional day was today and I'm in my feels.