#1 Issues with PubSub
Corrected the badge to pull our actual count. This is working fine but I believe the sidebar may need to be a stateful component as well because this component isn't getting events. I'm going to go that route but I at least wanted to include this just in case I need to revert.