Sidebar refactored to LiveView
Leaving the stateless component for the time being, just in case I'm missing something.
Fix titles to be uniform.
Notifications now require or at least use PubSub to communicate between the NotificationLive and the SidebarLive where the badge lives. We can send increment and decrement updates to prevent needing to query the database.