Layout adjustments
Move logo to sidebar to pull a complete Livebeats. There's no sense in having a header along the top tbh, not for just a logo and potentially a dark/light mode toggle.

Pulling a little sleight of hand to get notifications to push the footer down. Sticky doesn't mean along the bottom without having a height push it down. I can't for the life of me really work out a static sidebar, footer along the bottom, and scrolling text areas.

I did wrap the `<%= @inner_content %>` of `app.html.heex` with a div with `overflow-x-hidden overflow-y-auto w-full h-screen` and that made scrollable sections for each page but also kept a scrollbar along the outter section. We don't want 2 scrolls and we also technically don't want this behavior except for the artist index. That is the only "inner scrollbar" we need. I'm sure I could work out a more 1:1 with iTunes given time. I want to finish functionality at this point.