#1 Create a delivery stub
Add PubSub event to notification page to update the list of notifications as they come in. No query necessary though things will be out of order due to prepending to the head of a list. We can use Enum.reverse() but this event is called for every notification. It's a n+1 problem where every call reverses instead of waiting until the final notification is sent to reverse the whole list.