#4 Spotify rough draft
Fixed SpotifyEx by deleting the runtime config. I believe what was happening was runtime would set things up and the extra config would tear them down. It makes sense for everything to be in one place now.
Added `search_artist` function as a replacement. For some reason `Album.get_artists_albums` would error trying to build the list of structs when search is perfectly fine.

We are now transforming the data and I believe we could possibly go a step further. Upserts should be a case-insensitive comparison to work correctly because we're only capturing the name and release date. We could possibly parse the release date using the fields and Timex. The unique index is only on name anyway so this really doesn't matter except to have both fields being in sync. We can address that later anyway.