Broadcasting new albums with PubSub
Now we can track new album additions. This is in preparation for setting up badges as well as updating the LiveViews where new data could be visible. We want to update the UI when the backend backfills. It would be useful to also update whenever images change but this could happen for us.