#8 Oban verification working as expected \o/
I have no clue why I didn't just think of this before. There were edge cases with setting result here and while it's a little dirty, I think this is the best approach.

I know I can refactor getting the result of verify and calling `update_artist` would return an `(:ok, record}` tuple that wouldn't count as completion it felt like. I realized that if errors are thrown before the final :ok then it'll retry, which is exactly what we want but we can assume that if the code completed it shouldn't be retried. I think I knew this instinctively but it took way too long for it to click.