#19 Verification tests complete
Set untested functions to private. We really don't care to deeply test these and I'd say this test file is the most brittle and an integration test as it relies on Spotify communication.

I'm okay with the limited test to a single notification but that can be problematic in certain cases. I'm okay with the trade-offs of keeping the tests simple and promoting reuse. At first, I used identical test setups until I realized I needed to use `verify` and not build up the whole pipeline a second time.