#56 Dropdown component
Split out `verify` event name as it appears that the Sidebar wanted to take precedence or I just wasn't really creating unique patterns between the two.
I assumed phx-click would be possible but it's nice to know we can replicate our existing functionality though we do have an expectation of what goes in. If it's always going to be a flexbox expecting an icon then it should probably be listed as a slot prop. I think it's okay with what we have now though.